+++ RevStay/src/main/java/com/Rev/RevStay/services/BookingService.java
@@ -70,18 +70,17 @@ public class BookingService {
         // Logic to fetch room details (stubbed for now)
         return "Room details for " + roomType + " in " + hotel;
     }
-    public List<Booking> getBookingsByUser(Long userId) {
+    public List<Booking> getBookingsByUser(int userId) {
         return bookingDAO.findByUserId(userId);
     }
 
-    public Booking updateBookingStatus(Long bookingId, BookingStatus status, Integer userId) {
+    public Booking updateBookingStatus(int bookingId, BookingStatus status, Integer userId) {
         Optional<Booking> bookingOptional = bookingDAO.findById(bookingId);
 
         boolean validStatus = status.toString().equals("CANCELLED") || status.toString().equals("ACCEPTED");
         if(!validStatus) {
             throw new GenericException("Invalid status: " + status);
         }
-
 
         //Check if the user is the owner of the booking with the id in session
